netlify/functions/bot.py
#!/usr/bin/env python3
"""
SvetBot - Serverless version for Netlify Functions
Telegram Bot для мониторинга света в Киеве
"""

import json
import asyncio
import logging
import os
from datetime import datetime

# Telegram Bot imports
from telegram import Update
from telegram.ext import Application, CommandHandler, MessageHandler, filters, ContextTypes

# Веб-скрапинг imports
import requests
from bs4 import BeautifulSoup

# Настройка логирования
logging.basicConfig(
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    level=logging.INFO
)
logger = logging.getLogger(__name__)

# Конфигурация бота
class Config:
    BOT_TOKEN = os.getenv("BOT_TOKEN", "8362355096:AAGuP7hsn2Sg7QTJqrx76LqegJXBWBg-EbE")
    CHAT_ID = os.getenv("CHAT_ID", "-1002244805446")
    SITE_URL = "https://kyiv.energy-ua.info/grafik/Київ/вул.+Гміри+Бориса/14-А"
    CHECK_INTERVAL = 300  # 5 минут

config = Config()

def lambda_handler(event, context):
    """
    Serverless function handler for Netlify
    """
    try:
        # Webhook обработка
        if event.get('httpMethod') == 'POST':
            body = json.loads(event['body'])
            
            # Здесь будет обработка webhook от Telegram
            # Пока возвращаем успешный ответ
            return {
                'statusCode': 200,
                'headers': {
                    'Content-Type': 'application/json',
                    'Access-Control-Allow-Origin': '*'
                },
                'body': json.dumps({
                    'message': 'SvetBot webhook processed successfully',
                    'timestamp': datetime.now().isoformat()
                })
            }
        
        # GET запрос - статус бота
        if event.get('httpMethod') == 'GET':
            return {
                'statusCode': 200,
                'headers': {
                    'Content-Type': 'application/json',
                    'Access-Control-Allow-Origin': '*'
                },
                'body': json.dumps({
                    'status': 'active',
                    'bot_name': 'SvetBot',
                    'version': '2.0',
                    'features': ['light_monitoring', 'smoke_ranking', 'gifs', 'stickers'],
                    'timestamp': datetime.now().isoformat(),
                    'message': 'SvetBot is running on Netlify! 🤖⚡'
                })
            }
            
    except Exception as e:
        logger.error(f"Serverless function error: {e}")
        return {
            'statusCode': 500,
            'headers': {
                'Content-Type': 'application/json'
            },
            'body': json.dumps({
                'error': str(e),
                'timestamp': datetime.now().isoformat()
            })
        }

# Для локального тестирования
def handler(event, context):
    """Netlify Functions handler"""
    return lambda_handler(event, context)

if __name__ == "__main__":
    # Тестовое событие для локального запуска
    test_event = {
        'httpMethod': 'GET',
        'path': '/api/status'
    }
    test_context = {}
    
    result = lambda_handler(test_event, test_context)
    print(json.dumps(result, indent=2))