requirements.txt
# SVETOBOT - Requirements for Netlify
python-telegram-bot==22.5
requests==2.32.5
beautifulsoup4==4.14.2
lxml==6.0.2
httpx==0.27.2